+++ python/eyes_robotframework/src/EyesLibrary/base.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import, unicode_literals
 
+import contextlib
 from typing import TYPE_CHECKING, Text
 
 from AppiumLibrary import AppiumLibrary
@@ -10,6 +11,7 @@ from SeleniumLibrary import SeleniumLibrary
 from applitools.common.utils import cached_property
 from applitools.selenium import ClassicRunner, VisualGridRunner
 from applitools.selenium.fluent import SeleniumCheckSettings
+from applitools.selenium.runner import log_session_results_and_raise_exception
 
 from .config_parser import SelectedRunner
 from .errors import EyesLibraryValueError
@@ -30,7 +32,6 @@ class ContextAware(object):
         """Base class exposing attributes from the common context.
 
         :param ctx: The library itself as a context object.
-        :type ctx: SeleniumLibrary.SeleniumLibrary
         """
         self.ctx = ctx
         self.log = robot_logger
@@ -136,7 +137,7 @@ class LibraryComponent(ContextAware):
         # type: () -> Eyes
         return self.ctx.current_eyes
 
-    def _create_eyes_runner_if_needed(self):
+    def create_eyes_runner_if_needed(self):
         # type: () -> None
         if self.ctx.eyes_runner is None:
             # TODO: probably need to add runner_options to Configuration class
@@ -148,6 +149,18 @@ class LibraryComponent(ContextAware):
             self.ctx.eyes_runner = (
                 selected_runner(runner_options) if runner_options else selected_runner()
             )
+
+    @contextlib.contextmanager
+    def eyes_runner_get_all_test_results(self):
+        # type: () -> list[dict]
+        test_results = self.ctx.eyes_runner.get_all_test_results(False)
+        yield test_results
+        for r in test_results:
+            if r.exception is not None:
+                print("--- Test error. \n\tServer exception {}".format(r.exception))
+            else:
+                log_session_results_and_raise_exception(False, r.test_results)
+        self.ctx.clean_eyes_runner()
 
     def fetch_driver(self):
         # type: () -> AnyWebDriver