+++ python/eyes_robotframework/src/EyesLibrary/config.py
@@ -9,3 +9,4 @@ from applitools.selenium import RunnerOptions
 @attr.s
 class RobotConfiguration(Configuration):
     runner_options = attr.ib(default=None)  # type: Optional[RunnerOptions]
+    propagate_eyes_test_results = attr.ib(default=True)