+++ python/tests/functional/eyes_robotframework/test_integration.py
@@ -1,8 +1,14 @@
+import os
 import subprocess
 from os import path
-from textwrap import dedent
 
 import pytest
+from robot.result import ExecutionResult
+
+from EyesLibrary.test_results_manager import (
+    METADATA_EYES_TEST_RESULTS_URL_NAME,
+    METADATA_PATH_TO_EYES_RESULTS_NAME,
+)
 
 backend_to_backend_name = {
     "selenium": "SeleniumLibrary",
@@ -10,54 +16,57 @@ backend_to_backend_name = {
 }
 
 
-def run_robot(*args):
+def run_robot(*args, output_file_path=None):
     test_dir = path.join(path.dirname(__file__), "robot_tests")
-    call_args = ("python", "-m", "robot") + args
+    call_args = (
+        "python",
+        "-m",
+        "robot",
+        "--output={}.xml".format(output_file_path),
+        "--report={}.html".format(output_file_path),
+        "--log={}-log.html".format(output_file_path),
+    ) + args
     result = subprocess.run(
         call_args, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, cwd=test_dir
     )
     return result.returncode, result.stdout.decode()
 
 
-def lines(text):
-    return [line.strip() for line in dedent(text).strip().splitlines()]
-
-
-def test_suite_dir_classic_runner():
-    code, output = run_robot(
+@pytest.mark.parametrize("runner", ["web", "web_ufg"])
+@pytest.mark.parametrize(
+    "with_propagation",
+    [
+        [
+            "--prerebotmodifier",
+            "EyesLibrary.PropagateEyesTestResults",
+        ],
+        [],
+    ],
+    ids=lambda d: "manual propagate" if d else "auto patched",
+)
+def test_suite_dir_with_results_propagation_and_one_diff_in_report(
+    tmp_path, runner, with_propagation, local_chrome_driver
+):
+    output_file_path = os.path.join(tmp_path, "testsuite-dir")
+    args = with_propagation + [
         "--variable",
-        "RUNNER:web",
+        "RUNNER:{}".format(runner),
         "test_suite_dir",
-    )
-    expected = lines(
-        """
-        ==============================================================================
-        Test Suite Dir
-        ==============================================================================
-        Running test suite with `web` runner and `../applitools.yaml` config
-        Using library `SeleniumLibrary` as backend
-        Test Suite Dir.Suite1
-        ==============================================================================
-        Check Window Suite 1                                                  | PASS |
-        ------------------------------------------------------------------------------
-        Test Suite Dir.Suite1                                                 | PASS |
-        1 test, 1 passed, 0 failed
-        ==============================================================================
-        Test Suite Dir.Suite2
-        ==============================================================================
-        Check Window Suite 2                                                  | PASS |
-        ------------------------------------------------------------------------------
-        Test Suite Dir.Suite2                                                 | PASS |
-        1 test, 1 passed, 0 failed
-        ==============================================================================
-        Test Suite Dir.Suite3
-        ==============================================================================
-        Check Window Suite 3                                                  | PASS |
-        ------------------------------------------------------------------------------
-        """
-    )
-    assert lines(output)[: len(expected)] == expected
-    assert code == 0, output
+    ]
+    run_robot(*args, output_file_path=output_file_path)
+    result = ExecutionResult(output_file_path + ".xml")
+    not_passed = [
+        t for suite in result.suite.suites for t in suite.tests if t.status != "PASS"
+    ]
+    assert not_passed == []
+    assert METADATA_PATH_TO_EYES_RESULTS_NAME in result.suite.suites[0].metadata
+    assert METADATA_PATH_TO_EYES_RESULTS_NAME in result.suite.suites[1].metadata
+
+    # check that report html is containing 1 failed test
+    # need browser here to display js based content
+    local_chrome_driver.get("file://" + output_file_path + ".html")
+    assert METADATA_EYES_TEST_RESULTS_URL_NAME in local_chrome_driver.page_source
+    assert "1 test failed" in local_chrome_driver.page_source
 
 
 @pytest.mark.parametrize(
@@ -66,82 +75,51 @@ def test_suite_dir_classic_runner():
         ("web", "appium", "android"),
         ("web", "appium", "ios"),
         ("web", "selenium", "android"),
-        ("web", "selenium", "desktop"),
         ("web", "selenium", "ios"),
         ("web_ufg", "selenium", "desktop"),
     ],
     ids=lambda d: str(d),
 )
-def test_suite_web(data):
+def test_suite_web(data, tmp_path):
     runner, backend, platform = data
-    code, output = run_robot(
+    output_file_path = os.path.join(
+        tmp_path, "{}-{}-{}".format(runner, backend, platform)
+    )
+    run_robot(
         "--variablefile",
         "variables_test.py:{runner}:{backend}:{platform}".format(
             runner=runner, backend=backend, platform=platform
         ),
-        "--output={}-{}-{}.xml".format(runner, backend, platform),
-        "--log=NONE",
-        "--report=NONE",
         "web_only.robot",
+        output_file_path=output_file_path,
     )
-    expected = lines(
-        """
-        [ WARN ] No `config` set explicitly. Trying to find `applitools.yaml` in current path
-        Running test suite with `{runner}` runner and `applitools.yaml` config
-        Using library `{backend_name}` as backend
-        ==============================================================================
-        Web Only
-        ==============================================================================
-        Check Window                                                          | PASS |
-        ------------------------------------------------------------------------------
-        Check Window Fully                                                    | PASS |
-        ------------------------------------------------------------------------------
-        Check Region By Element                                               | PASS |
-        ------------------------------------------------------------------------------
-        Check Region By Selector                                              | PASS |
-        ------------------------------------------------------------------------------
-        Check Region By Selector With Ignore                                  | PASS |
-        ------------------------------------------------------------------------------
-        Check Window Two Times                                                | PASS |
-        ------------------------------------------------------------------------------
-        """.format(
-            runner=runner, backend_name=backend_to_backend_name[backend]
-        )
-    )
-    assert lines(output)[: len(expected)] == expected
-    assert code == 0, output
+    result = ExecutionResult(output_file_path + ".xml")
+    not_passed = [t for t in result.suite.tests if t.status != "PASS"]
+    assert not_passed == []
 
 
-def test_web_desktop_only():
+def test_web_desktop(
+    tmp_path,
+):
     runner, backend, platform = "web", "selenium", "desktop"
-    code, output = run_robot(
+    output_file_path = os.path.join(
+        tmp_path, "{}-{}-{}".format(runner, backend, platform)
+    )
+    run_robot(
         "--variablefile",
         "variables_test.py:{runner}:{backend}:{platform}".format(
             runner=runner, backend=backend, platform=platform
         ),
-        "--output={}-{}-{}.xml".format(runner, backend, platform),
-        "--log=NONE",
-        "--report=NONE",
         "web_desktop_only.robot",
+        "web_only.robot",
+        output_file_path=output_file_path,
     )
+    result = ExecutionResult(output_file_path + ".xml")
 
-    expected = lines(
-        """
-        [ WARN ] No `config` set explicitly. Trying to find `applitools.yaml` in current path
-        Running test suite with `{runner}` runner and `applitools.yaml` config
-        Using library `{backend_name}` as backend
-        ==============================================================================
-        Web Desktop Only
-        ==============================================================================
-        Check Shadow Dom                                                      | PASS |
-        ------------------------------------------------------------------------------
-        """.format(
-            runner=runner,
-            backend_name=backend_to_backend_name[backend],
-        )
-    )
-    assert lines(output)[: len(expected)] == expected
-    assert code == 0, output
+    not_passed = [
+        t for suite in result.suite.suites for t in suite.tests if t.status != "PASS"
+    ]
+    assert not_passed == []
 
 
 @pytest.mark.parametrize(
@@ -153,29 +131,21 @@ def test_web_desktop_only():
     ],
     ids=lambda d: str(d),
 )
-def test_suite_mobile_native(data):
+def test_suite_mobile_native(data, tmp_path):
     backend = "appium"
     platform, runner = data
-    code, output = run_robot(
+    output_file_path = os.path.join(
+        tmp_path, "{}-{}-{}".format(runner, backend, platform)
+    )
+    run_robot(
         "--variablefile",
         "variables_test.py:{runner}:{backend}:{platform}".format(
             runner=runner, backend=backend, platform=platform
         ),
         "mobile_native.robot",
+        output_file_path=output_file_path,
     )
-    expected = lines(
-        """
-        [ WARN ] No `config` set explicitly. Trying to find `applitools.yaml` in current path
-        Running test suite with `{runner}` runner and `applitools.yaml` config
-        Using library `{backend_name}` as backend
-        ==============================================================================
-        Mobile Native
-        ==============================================================================
-        Check Window Native                                                   | PASS |
-        ------------------------------------------------------------------------------
-        """.format(
-            runner=runner, backend_name=backend_to_backend_name[backend]
-        )
-    )
-    assert lines(output)[: len(expected)] == expected
-    assert code == 0, output
+
+    result = ExecutionResult(output_file_path + ".xml")
+    not_passed = [t for t in result.suite.tests if t.status != "PASS"]
+    assert not_passed == []